requirements.txt
PyYAML>=6.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
imbalanced-learn>=0.10.0
xgboost>=1.7.0
